% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{NVI.BoxPlot}
\alias{NVI.BoxPlot}
\title{Create Boxplot for the Best NVI Result}
\usage{
NVI.BoxPlot(res.NVI, filename = NULL)
}
\arguments{
\item{res.NVI}{A list object returned by the \code{analysis.NVI} function.}

\item{filename}{A base string for the output PDF file name.}
}
\value{
This function does not return a value. It saves a plot to a PDF file.
}
\description{
A wrapper function that generates a boxplot for the most stable
clustering solution identified by the \code{analysis.NVI} function.
}
\details{
This function simplifies the process of visualizing the most stable clustering profile.
It extracts the necessary components (\code{wh.best.NVI}, \code{Xclust.NVI}) from the result
object of \code{analysis.NVI} and passes them to the internal plotting function
\code{Xclust.boxplot.filename_new} to generate a labeled boxplot.
}
\seealso{
\code{\link{analysis.NVI}}, \code{\link{Xclust.boxplot.filename_new}}
}
