% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_CompReg.R
\name{stat}
\alias{stat}
\title{Compute Statistics for Compositional Regression}
\usage{
stat(X, y)
}
\arguments{
\item{X}{A numeric matrix of log-transformed compositional predictors.}

\item{y}{A numeric vector for the response variable.}
}
\value{
A list of key statistics and matrices required for inference.
}
\description{
An internal helper function to compute coefficients and t-statistics
for compositional regression. This is nearly identical to \code{CompReg}.
}
\keyword{internal}
